refactor to test backend endpoints
